+++ managers/webpubManifest.py
@@ -14,15 +14,12 @@ class WebpubManifest:
         self.openSection = None
 
     # TODO validate kwargs against schema.org/Book
-    def addMetadata(self, dcdwRecord: object, conformsTo=None) -> NoReturn:
+    def addMetadata(self, dcdwRecord: object, conformsTo=None) -> None:
         self.metadata['title'] = dcdwRecord.title
 
         if len(dcdwRecord.authors) > 0:
             print(dcdwRecord.authors)
-            if dcdwRecord.source == 'isac':
-                self.metadata['author'] = list(dcdwRecord.authors[0][0].split('|'))[0]    
-            else:
-                self.metadata['author'] = list(dcdwRecord.authors[0].split('|'))[0]
+            self.metadata['author'] = list(dcdwRecord.authors[0].split('|'))[0]
 
         isbns = list(filter(
             lambda x: x[-4:] == 'isbn', dcdwRecord.identifiers
@@ -51,7 +48,7 @@ class WebpubManifest:
         self.openSection = None
 
     # TODO Make it possible to add subsections iteratively
-    def addChapter(self, link: str, title: str, subsections=None) -> NoReturn:
+    def addChapter(self, link: str, title: str, subsections=None) -> None:
         component = {'href': link, 'title': title}
         tocEntry = component.copy()