+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Script to delete current duplicate authors/contributors in the PSQL database
 - Web scraper script for the Chicago ISAC catalog of publications
+- Mapping and process for Chicago ISAC catalog of publications
 ### Fixed
 - Modify agentParser method to reduce number of future duplicate agents
 - Install `wheel` with pip to fix fasttext build